+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Build: Add flatpak generation to github actions.
 * Assemby: An option to show assembly tab when there is no source file.
 * Signals: Add an window to show how GDB handles signals. And to set that behavior.
+* Macros: Add 10 macros for custom commands that are executed by clicking a button (or a hot key).
 
 ## [2.6] - 2025-10-07
 * Watchpoints: Fixed regression when adding commands to a watchpoint.